requirements.txt
aiohttp==1.0.5
async-timeout==2.0.0
beautifulsoup4==4.6.0
BiblioPixel==3.4.2
BiblioPixelAnimations==3.20180123.110953
BiblioPixelTriggers==1.0.1
certifi==2018.1.18
chardet==3.0.4
click==6.7
croniter==0.3.20
discord.py==0.16.12
Flask==0.12.2
gitdb2==2.0.3
GitPython==2.1.8
idna==2.6
itsdangerous==0.24
Jinja2==2.10
loady==1.7.8
MarkupSafe==1.0
multidict==4.0.0
pyserial==3.4
python-dateutil==2.6.1
PyYAML==3.12
requests==2.18.4
six==1.11.0
smmap2==2.0.3
spidev==3.2
urllib3==1.22
websockets==3.4
Werkzeug==0.14.1
